fix: 회원 오류 수정 - #17
